+++ KoAR.SaveEditor/App.xaml
@@ -1,14 +1,12 @@
 ﻿<Application x:Class="KoAR.SaveEditor.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:updates="clr-namespace:KoAR.SaveEditor.Updates"
     StartupUri="Views/Main/MainWindow.xaml">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
                 <ResourceDictionary Source="/PresentationFramework.Aero;component/Themes/Aero.NormalColor.xaml" />
             </ResourceDictionary.MergedDictionaries>
-            <updates:UpdateService x:Key="{x:Type updates:UpdateService}" />
         </ResourceDictionary>
     </Application.Resources>
 </Application>
