--- KoAR.SaveEditor/Updates/UpdateService.cs
@@ -1,80 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Net;
-using System.Threading;
-using System.Threading.Tasks;
-using KoAR.SaveEditor.Constructs;
-
-namespace KoAR.SaveEditor.Updates
-{
-    public sealed class UpdateService
-    {
-        private UpdateInfo? _update;
-
-        public event EventHandler<EventArgs<DownloadProgress>>? DownloadProgress;
-
-        public event EventHandler? UpdateChanged;
-
-        public UpdateInfo? Update
-        {
-            get => this._update;
-            private set
-            {
-                this._update = value;
-                this.UpdateChanged?.Invoke(this, EventArgs.Empty);
-            }
-        }
-
-        public async Task CheckForUpdatesAsync(CancellationToken cancellationToken = default)
-        {
-            IReleaseInfo[]? releases = await UpdateMethods.FetchUpdateReleasesAsync(cancellationToken: cancellationToken).ConfigureAwait(false);
-            this.Update = releases == null ? null : new UpdateInfo(releases);
-        }
-
-        public async Task DownloadUpdateAsync(string targetFileName, CancellationToken cancellationToken = default)
-        {
-            if (this.Update == null)
-            {
-                return;
-            }
-            const int interval = 250;
-            int bytesTransferred = 0, bytesPerInterval = 0;
-            using Timer timer = new Timer(OnTick, null, 0, interval);
-            try
-            {
-                HttpWebRequest request = WebRequest.CreateHttp(this.Update.Uri);
-                using WebResponse response = await request.GetResponseAsync().ConfigureAwait(false);
-                cancellationToken.ThrowIfCancellationRequested();
-                using Stream stream = response.GetResponseStream();
-                using FileStream fileStream = File.Create(targetFileName);
-                byte[] buffer = new byte[8192];
-                while (bytesTransferred < this.Update.Size)
-                {
-                    cancellationToken.ThrowIfCancellationRequested();
-                    int count = await stream.ReadAsync(buffer, 0, buffer.Length).ConfigureAwait(false);
-                    cancellationToken.ThrowIfCancellationRequested();
-                    await fileStream.WriteAsync(buffer, 0, count).ConfigureAwait(false);
-                    bytesTransferred += count;
-                    bytesPerInterval += count;
-                }
-                timer.Change(Timeout.Infinite, Timeout.Infinite);
-                this.OnDownloadProgress(new DownloadProgress(bytesTransferred, default));
-            }
-            catch (Exception e)
-            {
-                timer.Change(Timeout.Infinite, Timeout.Infinite);
-                this.OnDownloadProgress(e);
-                throw;
-            }
-
-            void OnTick(object _)
-            {
-                DownloadProgress progress = new DownloadProgress(bytesTransferred, bytesPerInterval * 1000d / interval);
-                bytesPerInterval = 0;
-                this.OnDownloadProgress(progress);
-            }
-        }
-
-        private void OnDownloadProgress(DownloadProgress data) => this.DownloadProgress?.Invoke(this, new EventArgs<DownloadProgress>(data));
-    }
-}
