+++ KoAR.SaveEditor/Views/Main/MainWindowViewModel.cs
@@ -29,6 +29,7 @@ namespace KoAR.SaveEditor.Views.Main
         public MainWindowViewModel()
         {
             this.CheckForUpdateCommand = new DelegateCommand(this.CheckForUpdate, () => !this._isCheckingForUpdate);
+            this.OpenUpdateWindowCommand = new DelegateCommand(() => this.OpenUpdateWindow());
             if (!(bool)DesignerProperties.IsInDesignModeProperty.GetMetadata(typeof(Window)).DefaultValue)
             {
                 Application.Current.Activated += this.Application_Activated;
@@ -73,7 +74,7 @@ namespace KoAR.SaveEditor.Views.Main
             set => this.SetValue(ref this._mode, value);
         }
 
-        public DelegateCommand OpenUpdateWindowCommand { get; } = new DelegateCommand(() => MainWindowViewModel.OpenUpdateWindow());
+        public DelegateCommand OpenUpdateWindowCommand { get; }
 
         public StashManagerViewModel? StashManager
         {
@@ -87,7 +88,7 @@ namespace KoAR.SaveEditor.Views.Main
             }
         }
 
-        public UpdateService UpdateService { get; } = (UpdateService)Application.Current.TryFindResource(typeof(UpdateService));
+        public UpdateNotifier UpdateNotifier { get; } = new UpdateNotifier();
 
         public void OpenFile()
         {
@@ -145,12 +146,6 @@ namespace KoAR.SaveEditor.Views.Main
             MessageBox.Show(Application.Current.MainWindow, $"Save successful! Original save backed up as {this.GameSave.FileName}.bak.", "KoAR Save Editor", MessageBoxButton.OK, MessageBoxImage.Information);
         }
 
-        private static bool OpenUpdateWindow()
-        {
-            UpdateWindow window = new UpdateWindow { Owner = Application.Current.MainWindow };
-            return window.ShowDialog().GetValueOrDefault();
-        }
-
         private async void Application_Activated(object sender, EventArgs e)
         {
             Application application = (Application)sender;
@@ -160,12 +155,12 @@ namespace KoAR.SaveEditor.Views.Main
             {
                 using CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
                 cancellationTokenSource.CancelAfter(2500);
-                await this.UpdateService.CheckForUpdatesAsync(cancellationTokenSource.Token).ConfigureAwait(false);
+                await this.UpdateNotifier.CheckForUpdatesAsync(cancellationTokenSource.Token).ConfigureAwait(false);
             }
             catch (OperationCanceledException)
             {
             }
-            if (Debugger.IsAttached || this.UpdateService.Update == null || !application.Dispatcher.Invoke(MainWindowViewModel.OpenUpdateWindow))
+            if (Debugger.IsAttached || this.UpdateNotifier.Update == null || !application.Dispatcher.Invoke(this.OpenUpdateWindow))
             {
                 await application.Dispatcher.InvokeAsync(this.OpenFile);
             }
@@ -212,7 +207,7 @@ namespace KoAR.SaveEditor.Views.Main
                 this.IsCheckingForUpdate = true;
                 using CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
                 cancellationTokenSource.CancelAfter(15000); // 15s
-                await this.UpdateService.CheckForUpdatesAsync(cancellationTokenSource.Token).ConfigureAwait(false);
+                await this.UpdateNotifier.CheckForUpdatesAsync(cancellationTokenSource.Token).ConfigureAwait(false);
             }
             catch
             {
@@ -221,9 +216,9 @@ namespace KoAR.SaveEditor.Views.Main
             finally
             {
                 this.IsCheckingForUpdate = false;
-                if (this.UpdateService.Update != null)
+                if (this.UpdateNotifier.Update != null)
                 {
-                    this._dispatcher.Invoke(MainWindowViewModel.OpenUpdateWindow);
+                    this._dispatcher.Invoke(this.OpenUpdateWindow);
                 }
             }
         }
@@ -235,6 +230,17 @@ namespace KoAR.SaveEditor.Views.Main
                 "Save before closing.\nFile will be saved and then the application will close.",
                 "Application will not close."
             );
+        }
+
+        private bool OpenUpdateWindow()
+        {
+            if (this.UpdateNotifier.Update == null)
+            {
+                return false;
+            }
+            using UpdateViewModel viewModel = new UpdateViewModel(this.UpdateNotifier.Update);
+            UpdateWindow window = new UpdateWindow { DataContext = viewModel, Owner = Application.Current.MainWindow };
+            return window.ShowDialog().GetValueOrDefault();
         }
     }
 }
