+++ KoAR.SaveEditor/Views/Updates/UpdateWindow.xaml
@@ -6,10 +6,9 @@
     xmlns:editor="clr-namespace:KoAR.SaveEditor"
     xmlns:local="clr-namespace:KoAR.SaveEditor.Views.Updates"
     xmlns:constructs="clr-namespace:KoAR.SaveEditor.Constructs"
-    mc:Ignorable="d" d:DataContext="{d:DesignInstance {x:Type local:UpdateViewModel}}"
-    Icon="/app.ico" Title="Update Available - KoAR Save Editor"
-    Width="640" Height="480" WindowStartupLocation="CenterOwner" ResizeMode="CanResizeWithGrip"
-    DataContext="{constructs:Activator {x:Type local:UpdateViewModel}}">
+    mc:Ignorable="d" d:DataContext="{d:DesignInstance {x:Type local:UpdateViewModelBase}}"
+    Icon="/app.ico" Title="{Binding Title, StringFormat='{}{0} - KoAR Save Editor'}"
+    Width="640" Height="480" WindowStartupLocation="CenterOwner" ResizeMode="CanResizeWithGrip">
     <Window.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
@@ -30,8 +29,8 @@
         </Grid.ColumnDefinitions>
         <TextBlock Grid.Row="0" Text="Current Version:" TextAlignment="Right" FontWeight="SemiBold" Margin="2" />
         <TextBlock Grid.Row="0" Grid.Column="1" Margin="2" Text="{Binding Source={x:Static editor:App.Version}, StringFormat='v{0}'}"/>
-        <TextBlock Grid.Row="1" Text="Update Version:" TextAlignment="Right" FontWeight="SemiBold" Margin="2" />
-        <TextBlock Grid.Row="1" Grid.Column="1" Margin="2" Text="{Binding UpdateService.Update.Version, StringFormat='v{0}'}" />
+        <TextBlock Grid.Row="1" Text="Target Version:" TextAlignment="Right" FontWeight="SemiBold" Margin="2" />
+        <TextBlock Grid.Row="1" Grid.Column="1" Margin="2" Text="{Binding Target.Version, StringFormat='v{0}'}" />
         <TextBlock Grid.Row="2" Text="Release Notes:" TextAlignment="Right" FontWeight="SemiBold" Margin="2" VerticalAlignment="Top" />
         <RichTextBox Grid.Row="2" Grid.Column="1" FocusVisualStyle="{x:Null}" IsReadOnly="True" Margin="2" VerticalScrollBarVisibility="Auto" Loaded="RichTextBox_Loaded" />
         <StackPanel Grid.Row="3" Grid.ColumnSpan="2">
@@ -54,7 +53,7 @@
                         </Style.Triggers>
                     </Style>
                 </StackPanel.Style>
-                <constructs:StatefulProgressBar Minimum="0" Maximum="{Binding UpdateService.Update.Size}" Height="24" Value="{Binding BytesTransferred, Mode=OneWay, TargetNullValue=0}">
+                <constructs:StatefulProgressBar Minimum="0" Maximum="{Binding Target.ZipFileSize}" Height="24" Value="{Binding BytesTransferred, Mode=OneWay, TargetNullValue=0}">
                     <constructs:StatefulProgressBar.Style>
                         <Style TargetType="{x:Type constructs:StatefulProgressBar}">
                             <Style.Triggers>
@@ -81,7 +80,7 @@
                                     <DataTrigger.Binding>
                                         <MultiBinding Converter="{StaticResource EqualityConverter}">
                                             <Binding Path="BytesTransferred" />
-                                            <Binding Path="UpdateService.Update.Size" />
+                                            <Binding Path="Target.ZipFileSize" />
                                         </MultiBinding>
                                     </DataTrigger.Binding>
                                     <Setter Property="Text" Value="Preparing to update..." />
@@ -91,7 +90,7 @@
                                 <Setter.Value>
                                     <MultiBinding StringFormat="Transferred: {0} out of {1}; Current Speed: {2}/s">
                                         <Binding Path="BytesTransferred" Converter="{StaticResource ByteGroupsConverter}" />
-                                        <Binding Path="UpdateService.Update.Size" Converter="{StaticResource ByteGroupsConverter}" />
+                                        <Binding Path="Target.ZipFileSize" Converter="{StaticResource ByteGroupsConverter}" />
                                         <Binding Path="Speed" Converter="{StaticResource ByteGroupsConverter}" />
                                     </MultiBinding>
                                 </Setter.Value>
