+++ KoAR.SaveEditor/Views/Updates/UpdateWindow.xaml.cs
@@ -10,25 +10,25 @@ namespace KoAR.SaveEditor.Views.Updates
 
         protected override void OnClosing(CancelEventArgs e)
         {
-            using UpdateViewModel viewModel = (UpdateViewModel)this.DataContext;
+            UpdateViewModelBase viewModel = (UpdateViewModelBase)this.DataContext;
             PropertyChangedEventManager.RemoveHandler(viewModel, this.ViewModel_DialogResultChanged, nameof(viewModel.DialogResult));
             base.OnClosing(e);
         }
 
         protected override void OnPropertyChanged(DependencyPropertyChangedEventArgs e)
         {
-            if (e.Property == FrameworkElement.DataContextProperty && e.NewValue is UpdateViewModel viewModel)
+            if (e.Property == FrameworkElement.DataContextProperty && e.NewValue is UpdateViewModelBase viewModel)
             {
                 PropertyChangedEventManager.AddHandler(viewModel, this.ViewModel_DialogResultChanged, nameof(viewModel.DialogResult));
             }
             base.OnPropertyChanged(e);
         }
 
-        private void RichTextBox_Loaded(object sender, RoutedEventArgs e) => ((RichTextBox)sender).Document = ((UpdateViewModel)this.DataContext).FlowDocument;
+        private void RichTextBox_Loaded(object sender, RoutedEventArgs e) => ((RichTextBox)sender).Document = ((UpdateViewModelBase)this.DataContext).Document;
 
         private void ViewModel_DialogResultChanged(object sender, PropertyChangedEventArgs e)
         {
-            this.DialogResult = ((UpdateViewModel)sender).DialogResult;
+            this.DialogResult = ((UpdateViewModelBase)sender).DialogResult;
             this.Close();
         }
     }
